+++ project/Sources/mcu.c
@@ -60,7 +60,7 @@ void config(void) {
     TPM_habilitaNVICIRQ(18, 3);  // TPM1
 }
 
-float get_time(void) {
+uint8_t get_time(void) {
     // TODO
     return 0;
 }
